dynamic bitcoind ports, remove serial, move BLS unit tests to separate file
